+++ test/test.sh
@@ -304,6 +304,24 @@ test_subcommand_help_does_not_need_ticket_store() {
     rm -rf "$dir"
 }
 
+test_create_reads_description_from_stdin() {
+    local dir id
+    dir=$(new_workspace)
+
+    run_in_dir "$dir" bash -c 'printf "%s\n" "Line one" "\`literal selector\`" "\$VALUE stays literal" | "$1" create "Stdin description" -d -' _ "$TK"
+    assert_status 0
+    id="$LAST_OUTPUT"
+
+    run_in_dir "$dir" "$TK" show "$id"
+    assert_status 0
+    assert_contains "# Stdin description"
+    assert_contains "Line one"
+    assert_contains '`literal selector`'
+    assert_contains '$VALUE stays literal'
+
+    rm -rf "$dir"
+}
+
 test_create_show_and_status_flow() {
     local dir id
     dir=$(new_workspace)
@@ -499,6 +517,7 @@ main() {
     run_test test_create_retries_on_id_collision
     run_test test_parser_errors_are_not_suppressed
     run_test test_subcommand_help_does_not_need_ticket_store
+    run_test test_create_reads_description_from_stdin
     run_test test_create_show_and_status_flow
     run_test test_dep_and_undep_flow
     run_test test_partial_id_resolution_and_ambiguity